linalg: scale-invariant rank detection + solve; fix rank-0 IndexError (both backends)
Repairs scale-sensitive rank handling in the shared OLS backend. A covariate on
a large raw scale (~1e8) inflated the pivoted-QR rank threshold (anchored to the
largest pivot diagonal) and false-dropped the intercept/treatment/interaction to
NaN on a full-rank design; the scipy.lstsq(cond=1e-7) solve likewise truncated
the small-scale direction, returning finite-but-wrong coefficients.

_detect_rank_deficiency now runs a raw pivoted QR first (preserving the
established drop-column selection for genuinely collinear designs) and only
adopts the equilibrated rank when the raw drop was scale-induced. The solve
equilibrates columns to unit 2-norm and unscales the coefficients. Mirrored in
rust/src/linalg.rs (unscale before fitted/vcov). Rank-0 designs now return
all-NaN cleanly (solve_ols) or raise a clear ValueError (solve_logit/poisson,
routed through CS pscore_fallback) instead of a cryptic IndexError.

No drop-order change for collinear designs; no-op for full-rank well-conditioned
designs (R-parity goldens unaffected). New regression tests in
tests/test_linalg.py::TestNumericalStability.

+++ diff_diff/staggered.py
@@ -1326,8 +1326,12 @@ class CallawaySantAnna(
                             stacklevel=2,
                         )
                     cho = None  # Force lstsq path for ALL rank-deficient cases
+                    # dtype=int so an empty result (rank 0) is an int index array;
+                    # X_ctrl always carries an intercept column (positive norm,
+                    # independent), so kept_cols.size >= 1 here in practice.
                     kept_cols = np.array(
-                        [i for i in range(X_ctrl.shape[1]) if i not in dropped_cols]
+                        [i for i in range(X_ctrl.shape[1]) if i not in dropped_cols],
+                        dtype=int,
                     )
                 else:
                     kept_cols = None  # Full rank — use all columns
